resolve inconsistencies
original complaint:

The provisions of Clause 7 in the Series Seed Investment Agreement are not consistent with article 116.1 “Board Composition” in the M&AA.
In Clause 7, only the Common Control Holders get to nominate the Common Board Designee.  This is in contrast with the M&AA which refers to all holders of Common Stock. Please amend to be consistent with the M&AA version.
In Clause 7,  the Series Seed Board Designee is nominated by the Purchasers. The M&AA refers to all holders of the Series Seed Preference Stock. Please amend to be consistent with the M&AA version.
In Clause 7, the Mutual Consent Board Designee is nominated by the two groups of shareholders, each group voting separately and requiring majority consent within the group.  In the M&AA, this third director is nominated by vote of all shareholders voting together as one group. Please amend to be consistent with Clause 7.