considerations for CPs and completion
Completion Date to be defined. And assuming that the conditions are objectively ascertainable / purely factual events. Otherwise, there would be (1) a gap in who is to determine when such condition is considered as having been met and (2) uncertainty as to how the other party (who is not the one deciding) is to know that this has been occurred, and when its obligations under the Agreement have been activated.